move all the data utils to individual files
