+++ src/utils/data/getAssociationDetails.js
@@ -1,0 +1,6 @@
+// @flow
+import { Map, get } from 'immutable';
+
+const getAssociationDetails = (neighbor :Map | Object) :Map | Object => get(neighbor, 'associationDetails');
+
+export { getAssociationDetails };
