+++ src/utils/data/getNeighborDetails.js
@@ -1,0 +1,6 @@
+// @flow
+import { Map, get } from 'immutable';
+
+const getNeighborDetails = (neighbor :Map | Object) :Map | Object => get(neighbor, 'neighborDetails');
+
+export { getNeighborDetails };
