+++ src/utils/data/index.js
@@ -1,9 +1,7 @@
 // @flow
-import * as EntityUtils from './EntityUtils';
-import * as NeighborUtils from './NeighborUtils';
 
-/* eslint-disable import/prefer-default-export */
-export {
-  EntityUtils,
-  NeighborUtils,
-};
+export * from './formatNewEntityData';
+export * from './getAssociationDetails';
+export * from './getAssociationESID';
+export * from './getNeighborDetails';
+export * from './getNeighborESID';
